simple changelog function
With the auto-deploy of 'chat-gui', which is all this will run for (for
now) will always be up to date locally. Open the git repo locally, get
the last 3 commits, mutilate them, then write to chat

The whitespace/newline stripping can probably be done better, but I was
lazy. And it might be best to cache recent results but this seems like a
stretch goal. I'm not picking or looking for any commit name in
particular so we will get all.

My idea is on merge of commit (squash on the PR), set a
descriptive but simple message and this will be used and not end up
obnoxious. These should be somewhat comprehensible to the average
chatter

The test repo is archived so I assume it is fine for testing, it would
be difficult for anything there to change. Welcome to a better idea
though.